pyproject.toml
[tool.poetry]
name = "bayesrul"
version = "0.2.0"
description = ""
authors = ["Arthur Viens <arthviens@gmail.com>, Luis Basora <luis.basora@yahoo.fr>"]

[tool.poetry.dependencies]
tyxe = { git = "https://github.com/TyXe-BDL/TyXe.git" }
python = ">=3.8,<3.11"
numpy = "^1.22.3"
pandas = "^1.4.2"
scikit-learn = "^1.0.2"
tqdm = "^4.64.0"
lmdb = "^1.3.0"
pytorch-lightning = "^1.6.1"
bayesian-torch = "^0.2.1"
ipykernel = "^6.13.0"
h5py = "^3.6.0"
matplotlib = "^3.5.1"
pyarrow = "^8.0.0"
jupyterlab = "^3.4.2"
ipywidgets = "^7.7.0"
statsmodels = "^0.13.2"
hydra-core = "^1.2.0"
uncertainty-toolbox = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
optuna = "^2.10.0"
optuna-dashboard = "^0.6.4"
plotly = "^5.8.0"
torchinfo = "^1.6.6"
seaborn = "^0.11.2"
altair = "^4.2.0"
pre-commit = "^2.13.0"
isort = "^5.9.1"
black = ">=21.6"
nb-black = "^1.0.7"
flake8 = "^3.9.2"
mypy = ">=0.950"
#wandb = "^0.13.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target_version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    __pycache__
  | \.git
  | \.mypy_cache
  | \.ipynb_checkpoints
  | build
  | dist
)/
'''

[tool.isort]
line_length = 80
profile = "black"
known_first_party = ["numpy", "pandas"]

[tool.mypy]
python_version = 3.8
platform = "posix"

color_output = true
pretty = true
show_column_numbers = true
strict = false

# probably redundant, but useful when switching strict to false
check_untyped_defs = true
ignore_missing_imports = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true